chore: update the secrets to use the existing GitHub PAT secret
When this was created by the dependabot tool it wasn't able to tell that
we already have a GitHub package repo PAT (personal access token)
configured. This commit changes the `token`s to use this secret.

The "auto-migrator" also decided that we have two private registries so
I am leaving that, but think we may be able to drop the
`https://npm.pkg.github.com`.

This commit is very similar to this commit in the `www` repo:
GoProperly/www@eaff489